+++ web/src/component/widget/integer_box_model.dart
@@ -33,6 +33,7 @@ class IntegerBoxModel extends mvp.ViewModel {
   String get code {
     return '''
 ui.VerticalPanel vPanel = new ui.VerticalPanel();
+vPanel.spacing = 4;
     
 vPanel.add(new ui.Label("Type for validation"));
 // Integer boxNan
@@ -70,6 +71,7 @@ String _validateField(ui.IntegerBox box) {
    */
   ui.Widget asWidget() {
     ui.VerticalPanel vPanel = new ui.VerticalPanel();
+    vPanel.spacing = 4;
     
     vPanel.add(new ui.Label("Type for validation"));
     // Integer boxNan