+++ web/src/component/widget/text_box_model.dart
@@ -34,25 +34,11 @@ A standard single-line text box.
   // Return code snipet how to use component
   String get code {
     return '''
-// Create a panel to layout the widgets
-ui.Grid grid = new ui.Grid(2, 2);
+ui.VerticalPanel panel = new ui.VerticalPanel();
+panel.add(new ui.Label("This is text box:"));
+panel.add(new ui.TextBox());
 
-grid.setWidget(0, 0, new ui.Label("Normal Text:"));
-// Add a normal and disabled text box
-ui.TextBox normalText = new ui.TextBox();
-// Set the normal text box to automatically adjust its direction according
-// to the input text. Use the Any-RTL heuristic, which sets an RTL direction
-// iff the text contains at least one RTL character.
-//normalText.enableDefaultDirectionEstimator(true);//setDirectionEstimator(AnyRtlDirectionEstimator.get());
-grid.setWidget(0, 1, normalText);
-
-grid.setWidget(1, 0, new ui.Label("Disabled Text:"));
-ui.TextBox disabledText = new ui.TextBox();
-disabledText.text = "read only";
-disabledText.enabled = false;
-grid.setWidget(1, 1, disabledText);
-
-return grid;
+return panel;
 ''';
   }
   
@@ -66,26 +52,9 @@ return grid;
    */
   ui.Widget asWidget() {
     
-    // Make some text boxes. The password text box is identical to the text
-    // box, except that the input is visually masked by the browser.
-    ui.PasswordTextBox ptb = new ui.PasswordTextBox();
-    ui.TextBox tb = new ui.TextBox();
-
-    // TODO(ECC) must be tested.
-    tb.addKeyPressHandler(new event.KeyPressHandlerAdapter((event.KeyPressEvent evt) {
-      print("${evt.getUnicodeCharCode()} : ${tb.getValue()}");
-    }));
-
-    // Let's make an 80x50 text area to go along with the other two.
-    ui.TextArea ta = new ui.TextArea();
-    ta.setCharacterWidth(80);
-    ta.setVisibleLines(50);
-
-    // Add them to the root panel.
     ui.VerticalPanel panel = new ui.VerticalPanel();
-    panel.add(tb);
-    panel.add(ptb);
-    panel.add(ta);
+    panel.add(new ui.Label("This is text box:"));
+    panel.add(new ui.TextBox());
     
     return panel;
   }